requirements.txt
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
pytest==8.1.1
pytest-asyncio==0.23.5
aioresponses==0.7.6
pytest-mock==3.12.0